chore: :truck: rename to use `update_` rather than `edit_`
