+++ src/seedcase_sprout/core/update_package_properties.py
